Update init.js
Removing the initReady hook in init.js as it is not loaded via loadScriptR API which sets the $script object used by the addHook object.  Now, the registerChatCommand() calls for chat, module and reconnect still function and are loaded:

[02-05-2016 @ 20:17:15] Loaded module: commands/topCommand.js (Enabled)
[02-05-2016 @ 20:17:15] Bot locked & loaded!
[02-05-2016 @ 20:17:15] Trying to add ./init.js | chat
[02-05-2016 @ 20:17:15] Trying to add ./init.js | module
[02-05-2016 @ 20:17:15] Trying to add ./init.js | reconnect

So, this is sort of a kludge, but, not entirely sure it is safe to try to build up a $script object before performing a $.bind('initReady') in init.js.